Release Version 2.0: Maestro and Logic Changes
Added support for Maestro card numbers.

The existing logic of using card length as the starting
conditional was inefficient and made detectNetwork
tough to understand visually. As it currently stands,
all card prefixes are unique. Becuase of this, it makes
the most sense to start our coniditional logic checking
across the prefixes first, and then confirm card type using
the card number length.