Support China UnionPay
Added support for Chila UnionPay prefixes.

Due to the wide range of prefixes for CUP,
it was necessary to build out a separate function to
confirm whether or not a card number contained a CUP prefix.

checkForChinaUnionPayPrefix() was crated, returning a true\false
value which is then leveraged by a conditional within detectNetwork().

Breaking out the other card prefixes into separate functions might be
worth considering for future developments.